pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "krunic"
version = "0.2.4"
description = "Tune image classifier using Ray Tune + SkyPilot"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argparse>=1.4.0",
    "gdown>=5.2.1",
    "h5py>=3.16.0",
    "matplotlib>=3.10.8",
    "medmnist>=3.0.2",
    "numpy>=2.4.3",
    "optuna>=4.7.0",
    "pillow>=12.1.1",
    "ray[train,tune]>=2.54.0",
    "scikit-learn>=1.8.0",
    "skypilot[aws,lambda]==0.12.0",
    "timm>=1.0.25",
    "torch>=2.0.0",
    "webdataset>=1.0.2",
]

[tool.hatch.build.targets.wheel]
include = ["krunic/*.py", "krunic/cluster_requirements.txt"]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
tunic = "krunic.tunic:main"
krunic = "krunic.krunic:main"
tunic-plotter = "krunic.tunic_plotter:main"
cvic = "krunic.cvic:main"
